neurodamus: move core modifications into packages
