+++ bluebrain/repo-bluebrain/packages/model-neocortex/package.py
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
 from spack import *
-from .sim_model import SimModel
+from .sim_model import SimModel, copy_all, make_link
 
 
 class ModelNeocortex(SimModel):
